move all equations to Equations class
